Refactor: extract keyboard and status line helpers
